consolidate common functions in process models
